+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.2.0
+
+- Added `dir()` method to return the target element's directionality
+- Added `lang()` method to return the target element's language
+
 ## 2.1.3
 
 - Renamed `updateLocalizedTerms()` to `update()` (forgive me SemVer, but nobody was using this I promise)